project/requirements.txt
fastapi==0.61.1
uvicorn==0.15.0
python-dotenv==0.19.0
PyYAML==5.4.1
pytest==6.2.5
requests==2.26.0
coverage==5.5
flake8==3.9.2


